Add idx translation for graph

+++ tests/graph/to_unified.rs
@@ -24,7 +24,7 @@ fn test_access_cached_after_unified_construct_buf() -> custos::Result<()> {
         .cache()
         .nodes
         .get(&Ident {
-            idx: cl_cpu_buf.ident.idx as usize,
+            idx: cl_cpu_buf.ident.idx,
             len: cl_cpu_buf.len(),
         })
         .unwrap()
